Popravil funkcijo

+++ model.py
@@ -1,4 +1,5 @@
 import json
+import os.path
 
 class Udelezenec:
     def __init__(self, ime, priimek) -> None:
@@ -8,10 +9,22 @@ class Udelezenec:
 
 def v_seznam(udelezenec):
     """"Objekt 'udeleženec' zapiše v seznam v JSON datoteki."""
-    with open('seznam.json', 'r') as dat:
-        seznam = json.load(dat)
-    
-    seznam.append(json.dumps(udelezenec.__dict__))
 
-    with open('seznam.json', 'w') as dat:
-        json.dump(seznam, dat, indent=4)
+# Ali datoteka s seznamom že obstaja? 
+
+    if os.path.isfile('seznam.json'):
+        with open('seznam.json', 'r') as dat:
+            seznam = json.load(dat)
+
+        seznam.append(json.dumps(udelezenec.__dict__))
+
+        with open('seznam.json', 'w') as dat:
+            json.dump(seznam, dat, indent=4)
+
+# Če ne obstaja, odpri novo in 'udelezenca' dodaj v prazen seznam.
+
+    else:
+        with open('seznam.json', 'a') as dat:
+            seznam = []
+            seznam.append(udelezenec.__dict__)
+            dat.write(json.dumps(seznam))
